# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_skiix-backend_internal_domain.DailyUsageHistoryItem:
    properties:
      date:
        type: string
      limit_reached:
        type: boolean
      posts_viewed:
        type: integer
      scroll_minutes:
        type: integer
    type: object
  github_com_skiix-backend_internal_domain.FeedLimitOverride:
    properties:
      activated_at:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      override_date:
        type: string
      user_id:
        type: string
    type: object
  github_com_skiix-backend_internal_domain.FeedLimitSetting:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_enabled:
        type: boolean
      max_overrides_per_day:
        description: default 3
        type: integer
      max_posts_per_day:
        description: NULL = unlimited
        type: integer
      max_scroll_minutes:
        description: NULL = unlimited
        type: integer
      override_duration_minutes:
        description: default 15
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_skiix-backend_internal_domain.FeedUsageResponse:
    properties:
      date:
        type: string
      is_limit_reached:
        type: boolean
      limit_reached_at:
        type: string
      limits:
        $ref: '#/definitions/github_com_skiix-backend_internal_domain.FeedLimitSetting'
      override:
        $ref: '#/definitions/github_com_skiix-backend_internal_domain.OverrideInfo'
      posts_viewed:
        type: integer
      progress:
        $ref: '#/definitions/github_com_skiix-backend_internal_domain.UsageProgress'
      scroll_minutes:
        type: integer
    type: object
  github_com_skiix-backend_internal_domain.FocusModeStatus:
    properties:
      blocked_categories:
        items:
          type: string
        type: array
      is_enabled:
        type: boolean
    type: object
  github_com_skiix-backend_internal_domain.OverrideInfo:
    properties:
      expires_at:
        type: string
      is_active:
        type: boolean
      overrides_remaining:
        type: integer
      overrides_used_today:
        type: integer
    type: object
...
  github_com_skiix-backend_internal_domain.TrackConsumptionRequest:
    properties:
      count:
        type: integer
      event_type:
        type: string
    type: object
  github_com_skiix-backend_internal_domain.TrackResponse:
    properties:
      can_override:
        type: boolean
      is_limit_reached:
        type: boolean
      message:
        type: string
      overrides_remaining:
        type: integer
      posts_remaining:
        type: integer
      posts_viewed:
        type: integer
      scroll_minutes:
        type: integer
      scroll_minutes_remaining:
        type: integer
    type: object
...
  github_com_skiix-backend_internal_domain.UpdateFeedLimitRequest:
    properties:
      is_enabled:
        type: boolean
      max_overrides_per_day:
        type: integer
      max_posts_per_day:
        type: integer
      max_scroll_minutes:
        type: integer
      override_duration_minutes:
        type: integer
    type: object
  github_com_skiix-backend_internal_domain.UsageAverages:
    properties:
      avg_posts_per_day:
        type: number
      avg_scroll_minutes:
        type: number
      days_limit_reached:
        type: integer
      total_days:
        type: integer
    type: object
  github_com_skiix-backend_internal_domain.UsageHistoryResponse:
    properties:
      averages:
        $ref: '#/definitions/github_com_skiix-backend_internal_domain.UsageAverages'
      history:
        items:
          $ref: '#/definitions/github_com_skiix-backend_internal_domain.DailyUsageHistoryItem'
        type: array
    type: object
  github_com_skiix-backend_internal_domain.UsageProgress:
    properties:
      posts_percentage:
        type: number
      scroll_percentage:
        type: number
    type: object
...
  internal_delivery_http.ErrorResponse:
    properties:
      error:
        example: invalid email format
        type: string
    type: object
...
  internal_delivery_http.SetCategoriesRequest:
    properties:
      categories:
        example:
        - entertainment
        - news
        - memes
        items:
          type: string
        type: array
    type: object
  internal_delivery_http.SetCategoriesResponse:
    properties:
      blocked_categories:
        example:
        - entertainment
        - news
        - memes
        items:
          type: string
        type: array
      message:
        example: blocked categories updated
        type: string
    type: object
  internal_delivery_http.ToggleRequest:
    properties:
      is_enabled:
        example: true
        type: boolean
    type: object
  internal_delivery_http.ToggleResponse:
    properties:
      is_enabled:
        example: true
        type: boolean
      message:
        example: focus mode updated
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Skiix authentication API with JWT and OAuth2 support
  title: Skiix API
  version: "1.0"
paths:
...
  /api/v1/feed-limit:
    delete:
      description: Disables the daily feed limit. Settings are preserved but the feature
        is turned off.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Disable feed limit
      tags:
      - feed-limit
    get:
      description: Retrieves the current daily feed limit configuration. Returns defaults
        if not configured.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_skiix-backend_internal_domain.FeedLimitSetting'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get feed limit settings
      tags:
      - feed-limit
    put:
      consumes:
      - application/json
      description: Creates or updates the user's daily feed limit configuration.
      parameters:
      - description: Feed limit settings
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_skiix-backend_internal_domain.UpdateFeedLimitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_skiix-backend_internal_domain.FeedLimitSetting'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update feed limit settings
      tags:
      - feed-limit
  /api/v1/feed-limit/history:
    get:
      description: Returns daily feed usage summaries for the past N days (default
        7, max 30).
      parameters:
      - description: Number of days to retrieve (1-30)
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_skiix-backend_internal_domain.UsageHistoryResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get feed usage history
      tags:
      - feed-limit
  /api/v1/feed-limit/override:
    post:
      description: Grants temporary extended access when the daily limit has been
        reached.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_skiix-backend_internal_domain.FeedLimitOverride'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Activate feed limit override (snooze)
      tags:
      - feed-limit
  /api/v1/feed-limit/track:
    post:
      consumes:
      - application/json
      description: Reports a feed consumption event (post view or scroll time). Returns
        current limit status.
      parameters:
      - description: Consumption event
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_skiix-backend_internal_domain.TrackConsumptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_skiix-backend_internal_domain.TrackResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Track feed consumption
      tags:
      - feed-limit
  /api/v1/feed-limit/usage:
    get:
      description: Returns the user's feed consumption stats for today versus their
        configured limits.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_skiix-backend_internal_domain.FeedUsageResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get today's feed usage
      tags:
      - feed-limit
...
  /focus-mode:
    get:
      consumes:
      - application/json
      description: Retrieve the current focus mode status and blocked categories for
        the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Focus mode status retrieved
          schema:
            $ref: '#/definitions/github_com_skiix-backend_internal_domain.FocusModeStatus'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - Bearer: []
      summary: Get focus mode status
      tags:
      - focus-mode
    put:
      consumes:
      - application/json
      description: Enable or disable focus mode for the authenticated user
      parameters:
      - description: Toggle payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http.ToggleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Focus mode toggled
          schema:
            $ref: '#/definitions/internal_delivery_http.ToggleResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - Bearer: []
      summary: Toggle focus mode
      tags:
      - focus-mode
  /focus-mode/categories:
    put:
      consumes:
      - application/json
      description: Replace the list of blocked content categories for the authenticated
        user
      parameters:
      - description: Categories payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_delivery_http.SetCategoriesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Categories updated
          schema:
            $ref: '#/definitions/internal_delivery_http.SetCategoriesResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_delivery_http.ErrorResponse'
      security:
      - Bearer: []
      summary: Set blocked categories
      tags:
      - focus-mode
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
    x-tokenname: Authorization
swagger: "2.0"